# cmd/relay/openapi.yaml
openapi: 3.1.0
info:
  title: Relay Webhook Delivery API
  description: |
    Relay is a webhook delivery system that guarantees delivery through
    idempotency keys, persistent storage, exponential backoff retries,
    and circuit breakers.

    ## Authentication
    All API requests require an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    API requests are rate limited per API key. Check response headers for limits:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets
  version: 1.0.0
  contact:
    name: Relay Support
    url: https://github.com/relay/relay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.relay.example.com
    description: Production
...
paths:
...
  /api/v1/events:
    post:
      tags:
        - Events
      summary: Create a webhook event
      description: |
        Creates a new webhook event for delivery. The event will be queued
        and delivered to the specified destination with automatic retries.

        Use the `X-Idempotency-Key` header to prevent duplicate events.
      operationId: createEvent
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Idempotency conflict - event already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    get:
      tags:
        - Events
      summary: List events
      description: Returns a paginated list of webhook events
      operationId: listEvents
      parameters:
        - name: status
          in: query
          description: Filter by event status
          schema:
            $ref: "#/components/schemas/EventStatus"
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Pagination cursor from previous response
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/events/{eventId}:
    get:
      tags:
        - Events
      summary: Get event by ID
      description: Returns a single event with its delivery history
      operationId: getEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithAttempts"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/events/{eventId}/replay:
    post:
      tags:
        - Events
      summary: Replay an event
      description: |
        Replays a failed or dead event. The event will be re-queued
        for delivery with a fresh retry count.
      operationId: replayEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event replayed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Event cannot be replayed (not in failed/dead status)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/events/batch/retry:
    post:
      tags:
        - Events
      summary: Batch retry events
      description: |
        Retry multiple failed or dead events in a single request.
        Supports three modes:
        - By IDs: Retry specific events by their UUIDs
        - By status: Retry all events with a given status (failed/dead)
        - By endpoint: Retry all failed/dead events for a specific endpoint
      operationId: batchRetryEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRetryRequest"
      responses:
        "200":
          description: Batch retry result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchRetryResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/stats:
    get:
      tags:
        - Health
      summary: Queue statistics
      description: Returns current queue and delivery statistics
      operationId: getStats
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/event-types:
    post:
      tags:
        - EventTypes
      summary: Create an event type
      description: |
        Creates a new event type with optional JSONSchema for payload validation.
        Event types allow you to define and document the structure of webhooks
        you send, and optionally validate payloads against the schema.
      operationId: createEventType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventTypeRequest"
      responses:
        "201":
          description: Event type created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventType"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Event type with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags:
        - EventTypes
      summary: List event types
      description: Returns a paginated list of event types
      operationId: listEventTypes
      parameters:
        - name: limit
          in: query
          description: Maximum number of event types to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of event types to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of event types
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTypeList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/event-types/{eventTypeId}:
    get:
      tags:
        - EventTypes
      summary: Get event type by ID
      description: Returns a single event type with its schema
      operationId: getEventType
      parameters:
        - $ref: "#/components/parameters/EventTypeId"
      responses:
        "200":
          description: Event type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventType"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - EventTypes
      summary: Update an event type
      description: Updates event type description and/or schema
      operationId: updateEventType
      parameters:
        - $ref: "#/components/parameters/EventTypeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventTypeRequest"
      responses:
        "200":
          description: Event type updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventType"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - EventTypes
      summary: Delete an event type
      description: Deletes an event type from the catalog
      operationId: deleteEventType
      parameters:
        - $ref: "#/components/parameters/EventTypeId"
      responses:
        "204":
          description: Event type deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/event-types/by-name/{eventTypeName}:
    get:
      tags:
        - EventTypes
      summary: Get event type by name
      description: Returns an event type by its name
      operationId: getEventTypeByName
      parameters:
        - name: eventTypeName
          in: path
          required: true
          description: Event type name (e.g., "order.created")
          schema:
            type: string
      responses:
        "200":
          description: Event type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventType"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # Alerting Endpoints
  /api/v1/alerting/rules:
    get:
      tags:
        - Alerting
      summary: List alert rules
      description: Returns all configured alert rules
      operationId: listAlertRules
      responses:
        "200":
          description: List of alert rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Alerting
      summary: Create an alert rule
      description: Creates a new alert rule for monitoring webhook delivery
      operationId: createAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRuleRequest"
      responses:
        "201":
          description: Alert rule created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/alerting/rules/{ruleId}:
    get:
      tags:
        - Alerting
      summary: Get alert rule by ID
      description: Returns a single alert rule
      operationId: getAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Alert rule details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Alerting
      summary: Update an alert rule
      description: Updates an existing alert rule
      operationId: updateAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertRuleRequest"
      responses:
        "200":
          description: Alert rule updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Alerting
      summary: Delete an alert rule
      description: Deletes an alert rule
      operationId: deleteAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Alert rule deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alerting/rules/{ruleId}/enable:
    post:
      tags:
        - Alerting
      summary: Enable an alert rule
      description: Enables a previously disabled alert rule
      operationId: enableAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Alert rule enabled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alerting/rules/{ruleId}/disable:
    post:
      tags:
        - Alerting
      summary: Disable an alert rule
      description: Disables an alert rule without deleting it
      operationId: disableAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Alert rule disabled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alerting/evaluate:
    post:
      tags:
        - Alerting
      summary: Evaluate alert rules
      description: Manually triggers evaluation of all enabled alert rules
      operationId: evaluateAlertRules
      responses:
        "200":
          description: Evaluation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  evaluated:
                    type: integer
                    description: Number of rules evaluated
                  triggered:
                    type: integer
                    description: Number of alerts triggered
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/alerting/history:
    get:
      tags:
        - Alerting
      summary: Get alert history
      description: Returns the history of triggered alerts
      operationId: getAlertHistory
      parameters:
        - name: limit
          in: query
          description: Maximum number of alerts to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Alert history
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Connector Endpoints
  /api/v1/connectors:
    get:
      tags:
        - Connectors
      summary: List connectors
      description: Returns all registered connectors
      operationId: listConnectors
      responses:
        "200":
          description: List of connectors
          content:
            application/json:
              schema:
                type: object
                properties:
                  connectors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Connector"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Connectors
      summary: Create a connector
      description: Creates a new connector for notifications or integrations
      operationId: createConnector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectorRequest"
      responses:
        "201":
          description: Connector created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Connector with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/connectors/{connectorName}:
    get:
      tags:
        - Connectors
      summary: Get connector by name
      description: Returns a single connector by its name
      operationId: getConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorName"
      responses:
        "200":
          description: Connector details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Connectors
      summary: Update a connector
      description: Updates an existing connector configuration
      operationId: updateConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnectorRequest"
      responses:
        "200":
          description: Connector updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Connectors
      summary: Delete a connector
      description: Deletes a connector
      operationId: deleteConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorName"
      responses:
        "204":
          description: Connector deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # Metrics Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      description: Event UUID
      schema:
        type: string
        format: uuid
...
    EventTypeId:
      name: eventTypeId
      in: path
      required: true
      description: Event type UUID
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      description: Unique key to prevent duplicate event creation
      schema:
        type: string
        minLength: 1
        maxLength: 255
    RuleId:
      name: ruleId
      in: path
      required: true
      description: Alert rule ID
      schema:
        type: string
    ConnectorName:
      name: connectorName
      in: path
      required: true
      description: Connector name
      schema:
        type: string
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        idempotencyKey:
          type: string
        destination:
          type: string
          format: uri
        eventType:
          type: string
        status:
          $ref: "#/components/schemas/EventStatus"
        attempts:
          type: integer
        maxAttempts:
          type: integer
        createdAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        nextAttemptAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - idempotencyKey
        - destination
        - status
        - attempts
        - maxAttempts
        - createdAt
    EventWithAttempts:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            payload:
              type: object
            headers:
              type: object
              additionalProperties:
                type: string
            deliveryAttempts:
              type: array
              items:
                $ref: "#/components/schemas/DeliveryAttempt"
...
    EventStatus:
      type: string
      enum:
        - queued
        - delivering
        - delivered
        - failed
        - dead
    BatchRetryRequest:
      type: object
      properties:
        event_ids:
          type: array
          items:
            type: string
            format: uuid
          description: List of event IDs to retry
        status:
          type: string
          enum: [failed, dead]
          description: Retry all events with this status
        endpoint_id:
          type: string
          format: uuid
          description: Retry all failed/dead events for this endpoint
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
          description: Maximum number of events to retry (for status/endpoint modes)
    BatchRetryResult:
      type: object
      properties:
        succeeded:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        failed:
          type: array
          items:
            $ref: "#/components/schemas/BatchRetryError"
        totalRequested:
          type: integer
        totalSucceeded:
          type: integer
      required:
        - succeeded
        - failed
        - totalRequested
        - totalSucceeded
...
    CreateEventRequest:
      type: object
      properties:
        destination:
          type: string
          format: uri
          description: Webhook destination URL (required if no eventType)
        eventType:
          type: string
          description: Event type for fan-out to subscribed endpoints
        payload:
          type: object
          description: JSON payload to deliver
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom headers to include in the webhook request
        maxAttempts:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum delivery attempts before marking as dead
      required:
        - payload
    EventList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
...
    QueueStats:
      type: object
      properties:
        queued:
          type: integer
        delivering:
          type: integer
        delivered:
          type: integer
        failed:
          type: integer
        dead:
          type: integer
        pending:
          type: integer
        processing:
          type: integer
        delayed:
          type: integer
    EventType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
        name:
          type: string
          description: Event type name (e.g., "order.created")
        description:
          type: string
          nullable: true
        schema:
          type: object
          nullable: true
          description: JSONSchema for payload validation
        schemaVersion:
          type: string
          nullable: true
          description: Schema version (e.g., "1.0")
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - name
        - createdAt
        - updatedAt
    CreateEventTypeRequest:
      type: object
      properties:
        name:
          type: string
          description: Event type name (e.g., "order.created")
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Human-readable description of this event type
        schema:
          type: object
          description: JSONSchema for payload validation (optional)
        schemaVersion:
          type: string
          description: Version of the schema (e.g., "1.0")
      required:
        - name
    UpdateEventTypeRequest:
      type: object
      properties:
        description:
          type: string
          description: Human-readable description of this event type
        schema:
          type: object
          nullable: true
          description: JSONSchema for payload validation (set to null to remove)
        schemaVersion:
          type: string
          description: Version of the schema
    EventTypeList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EventType"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
...
    AlertRule:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the alert rule
        name:
          type: string
          description: Human-readable name for the rule
        description:
          type: string
          description: Optional description of what the rule monitors
        enabled:
          type: boolean
          description: Whether the rule is actively being evaluated
        condition:
          $ref: "#/components/schemas/AlertCondition"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AlertAction"
          description: Actions to take when the alert fires
        cooldownMinutes:
          type: integer
          description: Minimum minutes between repeated alerts for the same rule
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - enabled
        - condition
        - actions
        - createdAt
        - updatedAt
...
    Alert:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the alert instance
        ruleId:
          type: string
          description: ID of the rule that triggered this alert
        ruleName:
          type: string
          description: Name of the rule that triggered this alert
        metric:
          $ref: "#/components/schemas/AlertMetric"
        value:
          type: number
          description: Actual metric value that triggered the alert
        threshold:
          type: number
          description: Threshold that was exceeded
        message:
          type: string
          description: Human-readable alert message
        triggeredAt:
          type: string
          format: date-time
          description: When the alert was triggered
      required:
        - id
        - ruleId
        - ruleName
        - metric
        - value
        - threshold
        - message
        - triggeredAt
    CreateAlertRuleRequest:
      type: object
      properties:
        name:
          type: string
          description: Human-readable name for the rule
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Optional description of what the rule monitors
        enabled:
          type: boolean
          default: true
        condition:
          $ref: "#/components/schemas/AlertCondition"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AlertAction"
          minItems: 1
        cooldownMinutes:
          type: integer
          minimum: 1
          default: 5
      required:
        - name
        - condition
        - actions
    UpdateAlertRuleRequest:
      type: object
      properties:
        name:
          type: string
          description: Human-readable name for the rule
        description:
          type: string
          description: Optional description of what the rule monitors
        enabled:
          type: boolean
        condition:
          $ref: "#/components/schemas/AlertCondition"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AlertAction"
        cooldownMinutes:
          type: integer
          minimum: 1

    # Connector Schemas
    Connector:
      type: object
      properties:
        name:
          type: string
          description: Unique name for the connector
        type:
          $ref: "#/components/schemas/ConnectorType"
        enabled:
          type: boolean
          description: Whether the connector is active
        config:
          $ref: "#/components/schemas/ConnectorConfig"
        templates:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ConnectorTemplate"
          description: Message templates keyed by notification type
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - type
        - enabled
        - config
        - createdAt
        - updatedAt
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
      required:
        - error
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid request body"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid API key"
            code: "UNAUTHORIZED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Event not found"
            code: "NOT_FOUND"
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "URL scheme must be http or https"
            code: "VALIDATION_ERROR"
...
security:
  - ApiKeyAuth: []